pyproject.toml
[project]
name = "prsbot"
version = "1.0.0"
description = "Production-ready Telegram bot for Russian infrastructure news monitoring"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.4",
    "httpx>=0.27",
    "tenacity>=8.2",
    "feedparser>=6.0",
    "beautifulsoup4>=4.12",
    "trafilatura>=1.8",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "PyYAML>=6.0",
    "SQLAlchemy>=2.0",
    "aiosqlite>=0.20",
    "APScheduler>=3.10",
    "simhash>=2.1",
    "structlog>=24.1",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
